Move JS symbol renaming to linkWasm
finalizedBy should be avoided:
* A finalizer execution order is not guaranteed.
In theory, a finalizer task can run at any time
after a finalized task.
In practise, Gradle runs a finalizer task
right after a finalized task (before all other tasks),
but there are no explicit guarantees of this behavior.

Also, modifying outputs of a finalized task in a finalizer
task breaks up-to-dateness of the finalized task.
In our case, `replaceSymbolsInSkikoJsOutput` modifies `linkWasm`
output, making it not up-to-date, so `linkWasm` is always re-run.

This commit moves .js file post-processing to linkWasm doLast action.