Revert customizing native artifacts coordinates
Customizing native publication's names does not work,
when a target cannot be compiled on a host.
A customization might lead to the inconsistent behavior,
when a kotlin multiplatform publication contains
different artifacts names, depending on a host system.
E.g. customization breaks macOS artifacts links,
if a kotlin multiplatform publication is built on Linux.
https://youtrack.jetbrains.com/issue/KT-50001