Remove Angle support (it is not fully implemented)
Angle is a stable Skia renderer, which is used by Chromium instead of direct DirectX, OpenGl, Metal renderers. It is considered more stable, but needs the ~30mb binary library. Angle mimics OpengGL ES and calls corresponding DirectX/Metal API's.

Angle support wasn't fully implemented, there just a bunch of class prototypes in the source code, that gets in the way during refactorings.

Deleting it for now, until we decide to fully support it. When we decide it, we can look at git history.