SkiaLayerAnalytics - send events for different steps of rendering (#568)
* SkiaLayerAnalytics - send events for different steps of rendering

It will help us to detect GPU's, on which we have crashes. After we compare two events, we can investigate GPU, and if the issue only in it, we will add it to not-supported-adapter.list

The plan to integrate this analytics to Toolbox.

* Fix build

* Disable performance tests on CI

CI can have different performance for different runs. Currently we have failing vsync tests for Windows

* Fix analytics test on github actions

* Fix CI