Expose SkiaLayer.redrawImmediately
Sometimes we need to update the current frame as fast as we can to avoid input lag.

For example, when we typed a character, or scrolled a list. If we render the new state only on the next frame, the user will experience input lag:
```
event -> change state -> needRedraw -> vsync -> draw -> vsync
```
If we would call this:
```
event -> change state -> redrawImmediately -> vsync
```
then there is no input lag.

We can't know in what condition we need to call it. The caller of SkiaLayer should decide it by itself. Usually it should be called when we change some state in response to an input event.