Windows. Fix a crash when we resize to zero and then to non-zero
Fixes https://github.com/JetBrains/compose-multiplatform/issues/4425

It is a regression after https://github.com/JetBrains/skiko/pull/858

The crash was because `d3dDevice->swapChain->GetBuffer` didn't return the buffer to draw on if we reused the buffer from the previous frame (we didn't change `surface` in case of zero size).

I am not completely sure why, but I exhausted my the investigation limit
and this fix is needed by other reasons (we need vsync on swap buffers).

## Testing
An additional check in the existed test (fails before the fix)