Add a flag to control number of buffers used by Metal. Add a flag to control Vsync on window resize (#968)
#### Number of buffers
MacOS
[doc](https://developer.apple.com/documentation/quartzcore/cametallayer/2938720-maximumdrawablecount?language=objc).
As far as I know AWT uses double buffering by default. So we would like
to experiment with this setting in Fleet. Also I expect that double
buffering should reduce user interaction latency.

#### Control VSync on window resize

I've noticed that on Linux and Windows we don't wait for VSync when
repainting synchronously, but only on macOS we do.
I suspect that locking EDT when we are waiting for VSync might lead to
problems with app responsivenes. So it would be nice to experiment with
this setting.