Remove BreakIterator.clone() method (#1024)
This PR is split-off from and is a prerequisite of Clang-cl PR: #1020

BreakIterator.clone() is broken on JVM and seemingly leads to a
use-after-free behaviour. Clang-cl makes this issue more visible, likely
due to address sanitisation.

Explanation why it is deleted in an internal channel:
https://jetbrains.slack.com/archives/C02DDNREC77/p1738678000232729
* `breakIteratorCloneTest` is failing with an access violation in Debug
mode with additional checks, so using it wasn't safe
* it is not used anywhere

This is a breaking change and will require bumping Skiko version to
0.9.0.