Make configureSignAndPublishDependencies use proper android task names (#1053)
"AndroidRelease" should be used instead of "Android" since using the
android plugin introduces release/debug variants.

Note: `./gradlew publish -Pskiko.android.enabled=true` would still fail
with the following, but this task is as I understand not used on CI and
"build repo" seem to be a local thing.
```
> Task :publishAndroidReleasePublicationToBuildRepoRepository FAILED

FAILURE: Build failed with an exception.

* What went wrong:
A problem was found with the configuration of task ':publishAndroidReleasePublicationToBuildRepoRepository' (type 'PublishToMavenRepository').
  - Gradle detected a problem with the following location: '/Users/cdelabou/workspace/misc/skiko/skiko/build/libs/skiko-0.0.0-SNAPSHOT-javadoc.jar.asc'.
    
    Reason: Task ':publishAndroidReleasePublicationToBuildRepoRepository' uses this output of task ':signKotlinMultiplatformPublication' without declaring an explicit or implicit dependency. This can lead to incorrect results being produced, depending on what order the tasks are executed.
```